convert to long long

+++ 2025/Quest 4/1.cpp
@@ -15,5 +15,5 @@ int main() {
         
     }
 
-    std::cout << ans * 2025;
+    std::cout << (long long)(ans * 2025);
 }
